Gather schema details from all nodes
This fixes validating against schemas that are included in child nodes,
and not defined in the root node.

Previously, the validator gathered schema details based only on the nsmap
for the root node.  Any namespace/schema dependencies defined in child
nodes are not discovered, so validation fails when it should not.

Example:

<STIX_Package (base ns/schema dependencies) >
   <Related_Packages>
       <Related_Package>
           <Package (additional child specific ns/schema dependencies) >
       </Related_Package>
   </Releated_Packages>
</STIX_Package>

Since only the nsmap of the root node is scanned to compile the schema
list, the child will fail to validate since its schemas will not be
available.